Introduce XDG for config and data dirs
In order to configure the application and to provide a place to store
catalogue data about the things that are being downloaded, we need to
find out where the best place on the system to store files is.

This commit pulls in the XDG crate so that we can use the Free Desktop
specification.

TODO: I have no idea at this point whether any of this is tenable on
Windows.