Flesh out methods for opening/bootstrapping the catalogue DB
This commit introdes a catalogue module. The intent of this module is
to provide an interface on top of a SQLite database that can be used
to keep track of which pages have been downloaded, which series they
belong to and other information about the manga that will allow me to
separate meta-information about manga from the actual image files on
disk.

The implementation is empty at the moment, the most interesting part
of this commit is that it introduces an OtakuError for the first
time. This is a custom Error Object for this application that we can
coerce other errors into to allow us to do proper error handling
within functions instead of relying on unwrap/except etc all the time.

I'm taking the approach that objects should only (unless absolutely
necessary) be unwrapped within the main function of a binary.