Convert library calls to expect into OtakuError
This commit goes back and converts functions that use expect and
unwrap and return raw types into functions that return Result<T,
OtakuError> using the error module defined in the previous
commit. This allows us to use the ? operator to wrap convert and
return wrapped errors into the OtakuError return type.

Note that all the impl's to convert From other error types basically
don't do anything at the moment, instead rely on the correct message
being set at the expect call in main, where the original error is
being passed up to. This is a potential refactoring in the future to
make the error type more intelligent but it'll do for now.

Also I'd like to find a way of not having to write all the boiler
plate From implementations, but again, this will do for now.