Bootstrap the database with a basic images table
When we download an image we want to write it to the Database so that
we know it's been downloaded and we can attach it to the relevant
Manga.

This commit is a first step in setting up the schema on application
boot from some sql files in src/bootstrap.