Return just the image name from the downloader
Previously we were returning the whole path, which would then get
saved in the database, returning just the image name allows us to move
the archive whereever in the filesystem that we want (potentially).

It also avoids having a load of redundant paths in the DB considering
that we're always saving data in XDG directories.