Move bootstrap check logic into the DB.
I like this approach much better, rather than just checking the schema
version and making an assumption, we encode a more robust set of
criteria for determining the state of the catalogue DB in some SQL
that returns 1 if the DB is fine and 0 otherwise.

That way we can map any errors or 0 results into an OtakuError and
return Result<(), OtakuError> from our function. The function logic
becomes simpler and we can use is_ok() at the callsite, which I like.