Place downloaded images in a standard location.
Previously in the tests and in main.rs I was just fudging /tmp into a
&str and passing that to the downloader as a place to put images.

Because the purpose of this application is to catalogue and keep all
the images that we've downloaded we should probably put them somewhere
specific. We can use our XDG data directories for that.

This commit makes sure that a catalogue directory exists inside our
XDG data dir and passes that to the downloader.

TODO:
* Should the download path and the catalogue db be combined to one
  coherant catalogue?