Remove the Catalogue struct
Doing this kind of OO style in Rust is not very idiomatic.

I'm struggling a little bit to write idiomatic Rust, however after
reading the chapter about OO in The Book, I'm convinced that this was
not how structs and traits are intended to work.

I'm in a position where I want to keep hold of an open database
connection for the duration of some processing that will involve
multiple queries and multiple calls to the DB.

My initial implementation involved creating a struct/impl pair for the
Catalogue that would function like a class in Ruby. I could hold the
open connection inside each 'instance' of the struct as an ivar and then
run methods on the struct-object that refer to the stored state.

I initially thought about creating a private function in the catalogue
that would open the database connection as required, but that adds much
more overhead to each function call as every time I do something with
the catalogue module I'd be opening a new database connection.

I'm not yet sure how best to implement this but for now I've settled for
flattening down the methods on the struct into functions attached to the
module, and passing a reference to the database connection as the
initial argument to each function