rename catalogue::is_bootstrapped -> db_state
This method doesn't return a bool so the is_ prefix isn't right. It
returns a result that is Ok if the db bootstrap script has succeeded
and Err if it's failed - so in this case the name db_state makes more
sense.

It also requires us to call is_ok() when using it in a branch
statement and

db_state().is_ok()

reads much better than

is_bootstrapped().is_ok()