Create a wrapper function in the catalogue image_to_catalogue
This function wraps the downloading and saving functionality to allow
us a single point at which to download an image and insert it into the
catalogue.

I toyed about with where to put this, but to me it feels much more
natural to call the downloader from within the catalogue rather than
to give the downloader access to a SQL connection.

This potentially means that as we now have a data path and a sql
connection to keep hold of that there is merit in returning some kind
of struct to represent the catalogue?